# product-catalog/api/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: FlexiShop Product Service API
  version: '1.0.0'
  description: API specification for FlexiShop's Product Service, incorporating best practices for pagination, error handling, and resource modeling. Uses individual query parameters for filtering to ensure specific URL format.
servers:
  - url: https://api.flexishop.com/v1
    description: Production server
paths:
  /products:
    get:
      tags:
        - Products
      summary: Retrieve a list of products
      description: Fetch a paginated list of products with support for filtering, sorting, projection, and inclusion of related resources using individual query parameters.
      operationId: listProducts
      parameters:
        # Pagination & Sorting
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/SizeParam'
        - $ref: '#/components/parameters/SortParam'
        # Filtering (Individual Parameters)
        - $ref: '#/components/parameters/SearchQueryParam' # q=...
        - $ref: '#/components/parameters/FilterByBrandParam' # brand=BrandName&brand=AnotherBrand
        - $ref: '#/components/parameters/FilterByPriceParam'   # price=gte:10&price=lte:50
        - $ref: '#/components/parameters/FilterByRatingParam'  # rating=gte:4
        - $ref: '#/components/parameters/FilterByProductCategoryParam' # product[category]=Electronics
        - $ref: '#/components/parameters/FilterByProductTagParam'      # product[tag]=wireless
        # Inclusion & Projection
        - $ref: '#/components/parameters/IncludeParam' # include=brand,tags
        - $ref: '#/components/parameters/FieldsParam'   # fields=id,name
      x-spring-provide-args:
        - 'com.gelerion.flexi.shop.product.catalog.infra.web.domain.FilterCriteria filterCriteria'
        - 'org.springframework.data.domain.Pageable pageable'
      responses:
        '200':
          description: A paginated list of products.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductPage'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'
    post:
      tags:
        - Products
      summary: Create a new product
      operationId: createProduct
      requestBody:
        description: Product data for creation.
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductCreateRequest'
      responses:
        '201':
          description: Product created successfully.
          headers:
            Location:
              description: URL of the created product.
              schema:
                type: string
                format: uri
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductResource'
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /products/{productId}:
    parameters:
      - $ref: '#/components/parameters/ProductIdParam'
    get:
      tags:
        - Products
      summary: Retrieve a product by ID
      description: Fetch a single product by its UUID, with optional inclusion of related resources and field projection.
      operationId: getProductById
      parameters:
        - $ref: '#/components/parameters/IncludeParam'
        - $ref: '#/components/parameters/FieldsParam'
      responses:
        '200':
          description: Product details. The included related resources depend on the 'include' parameter.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductResource'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
    patch:
      tags:
        - Products
      summary: Partially update an existing product
      description: Update one or more fields of an existing product. Only the fields provided in the request body will be updated.
      operationId: updateProduct
      requestBody:
        description: Fields to update in the product. All fields are optional.
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductUpdateRequest'
      responses:
        '200':
          description: Product updated successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductResource'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
        '500':
          $ref: '#/components/responses/InternalServerError'
    delete:
      tags:
        - Products
      summary: Delete a product
      operationId: deleteProduct
      responses:
        '204':
          description: Product deleted successfully.
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
components:
  parameters:
...
    ProductIdParam:
      name: productId
      in: path
      required: true
      schema:
        type: string
        format: uuid
      description: The UUID identifier of the product.
...
    PageParam:
      name: page
      in: query
      description: Page number of the requested page (0-indexed).
      required: false
      schema:
        type: integer
        default: 0
        minimum: 0
    SizeParam:
      name: size
      in: query
      description: Size of the requested page (number of items).
      required: false
      schema:
        type: integer
        default: 20
        minimum: 1
        maximum: 100
    SortParam:
      name: sort
      in: query
      description: |
        Sorting criteria in the format: `property,(asc|desc)`.
        Default sort order is ascending.
        Multiple sort criteria are supported (repeat the parameter).
        Example: `sort=name,asc&sort=price,desc`
      required: false
      schema:
        type: array
        items:
          type: string
      style: form
      explode: true # Allows multiple sort=... parameters

    # --- Filtering Parameters (Individual for GET /products) ---
    SearchQueryParam:
      name: q
      in: query
      description: Full-text search query string across relevant fields.
      required: false
      schema:
        type: string
    FilterByBrandParam: # Changed from ID to Name
      name: brand
      in: query
      description: Filters by one or more brand names. Repeat parameter for multiple values (e.g., `brand=Logitech&brand=Razer`).
      required: false
      schema:
        type: array
        items:
          type: string # Filter by name
      style: form
      explode: true # Generates brand=Logitech&brand=Razer
    FilterByPriceParam:
      name: price
      in: query
      description: |
        Filter by price range using `operator:value` format (e.g., `gte:10`, `lte:50`).
        Supported operators: `eq`, `gt`, `lt`, `gte`, `lte`.
        Provide one value for exact match/comparison or two values for a range (repeat parameter: `price=gte:10&price=lte:50`).
      required: false
      schema:
        type: array
        items:
          type: string
          pattern: '^(eq|gt|lt|gte|lte):(\d+(\.\d+)?)$' # Lowercase operators
        maxItems: 2
      style: form
      explode: true # Generates price=gte:10&price=lte:50
    FilterByRatingParam:
      name: rating
      in: query
      description: |
        Filter by rating range using `operator:value` format (e.g., `gte:3`, `lte:5`).
        Supported operators: `eq`, `gt`, `lt`, `gte`, `lte`.
        Provide one value for exact match/comparison or two values for a range (repeat parameter: `rating=gte:3&rating=lte:5`).
      required: false
      schema:
        type: array
        items:
          type: string
          pattern: '^(eq|gt|lt|gte|lte):(\d+)$' # Lowercase operators
        maxItems: 2
      style: form
      explode: true # Generates rating=gte:3&rating=lte:5
    FilterByProductCategoryParam: # Changed from ID to Name
      name: product[category]
      in: query
      description: Filters by one or more product category names. Repeat parameter for multiple values (e.g., `product[category]=Electronics&product[category]=Accessories`).
      required: false
      schema:
        type: array
        items:
          type: string # Filter by name
      style: form
      explode: true # Generates product[category]=Electronics&product[category]=Accessories
    FilterByProductTagParam: # Changed from ID to Name
      name: product[tag]
      in: query
      description: Filters by one or more product tag names. Repeat parameter for multiple values (e.g., `product[tag]=wireless&product[tag]=ergonomic`).
      required: false
      schema:
        type: array
        items:
          type: string # Filter by name
      style: form
      explode: true # Generates product[tag]=wireless&product[tag]=ergonomic

    # --- Inclusion & Projection Parameters ---
    IncludeParam:
      name: include
      in: query
      description: >
        Comma-separated list of related resources to include in the response.
        Controls which optional nested objects are present in the `ProductResource`.
        Possible values: `brand`, `tags`, `categories`, `images`, `specifications`.
      required: false
      schema:
        type: array
        items:
          type: string
          enum:
            - brand
            - tags
            - categories
            - images
            - specifications
      style: form
      explode: false # Generates include=brand,tags
    FieldsParam:
      name: fields
      in: query
      description: >
        Comma-separated list of top-level fields to include in the response resource(s).
        Applies to the main resource and any included related resources.
        Nested field selection (e.g., `brand.name`) is not supported via this parameter.
      required: false
      schema:
        type: array
        items:
          type: string
      style: form
      explode: false # Generates fields=id,name
  schemas:
...
    ProductResource:
      type: object
      description: Represents a product, potentially including related resources based on the 'include' parameter.
      required:
        - name
        - description
        - price
        - status
        - inventoryQuantity
      properties:
        id:
          type: string
          format: uuid
          readOnly: true
          description: Unique identifier for the product.
          example: '123e4567-e89b-12d3-a456-426614174000'
        name:
          type: string
          maxLength: 255
          description: Name of the product.
          example: 'Wireless Mouse'
          x-field-extra-annotation: "@jakarta.validation.constraints.NotNull\n@jakarta.validation.constraints.Size(min=1, max=255)"
        description:
          type: string
          description: Detailed description of the product.
          example: 'A high-quality wireless mouse with ergonomic design.'
          x-field-extra-annotation: "@jakarta.validation.constraints.NotNull"
        price:
          type: number
          format: float
          minimum: 0.0
          description: Price of the product.
          example: 29.99
          x-field-extra-annotation: "@jakarta.validation.constraints.NotNull"
        status:
          type: string
          enum: [active, inactive, discontinued]
          default: active
          description: Current status of the product.
          example: 'active'
        inventoryQuantity:
          type: integer
          minimum: 0
          default: 0
          description: Quantity currently in stock.
          example: 150
        createdAt:
          type: string
          format: date-time
          readOnly: true
          description: Timestamp when the product was created.
          example: '2023-10-01T12:00:00Z'
        updatedAt:
          type: string
          format: date-time
          readOnly: true
          description: Timestamp when the product was last updated.
          example: '2023-10-02T12:00:00Z'
        # Optional Related Resources (Removed nullable: true)
        brand:
          $ref: '#/components/schemas/BrandResource'
          description: Brand details (included if 'brand' is specified in 'include' parameter).
          readOnly: true
        categories:
          type: array
          items:
            $ref: '#/components/schemas/CategoryResource'
          description: List of categories (included if 'categories' is specified in 'include' parameter).
          readOnly: true
        tags:
          type: array
          items:
            $ref: '#/components/schemas/TagResource'
          description: List of tags (included if 'tags' is specified in 'include' parameter).
          readOnly: true
        images:
          type: array
          items:
            $ref: '#/components/schemas/ImageResource'
          description: List of product images (controlled by 'include=images').
          readOnly: true
        specifications:
          type: array
          items:
            $ref: '#/components/schemas/SpecificationResource'
          description: List of product specifications (controlled by 'include=specifications').
          readOnly: true
    BrandResource:
      type: object
      description: Represents a brand.
      required:
        - name
      properties:
        id:
          type: integer
          readOnly: true
          example: 2
        name:
          type: string
          maxLength: 255
          example: 'Logitech'
        description:
          type: string
          # Removed nullable: true, optionality handled by not being required
          example: 'A leading manufacturer of computer peripherals.'
        createdAt:
          type: string
          format: date-time
          readOnly: true
        updatedAt:
          type: string
          format: date-time
          readOnly: true
    CategoryResource:
      type: object
      description: Represents a product category.
      required:
        - name
      properties:
        id:
          type: integer
          readOnly: true
          example: 3
        name:
          type: string
          maxLength: 255
          example: 'Electronics'
          x-field-extra-annotation: "@jakarta.validation.constraints.NotNull\n@jakarta.validation.constraints.Size(min=1, max=255)"
        parentCategoryId:
          type: integer
          # Removed nullable: true
          example: 1
        createdAt:
          type: string
          format: date-time
          readOnly: true
        updatedAt:
          type: string
          format: date-time
          readOnly: true
    TagResource:
      type: object
      description: Represents a tag that can be applied to products.
      required:
        - name
      properties:
        id:
          type: integer
          readOnly: true
          example: 4
        name:
          type: string
          maxLength: 255
          example: 'wireless'
        createdAt:
          type: string
          format: date-time
          readOnly: true
        updatedAt:
          type: string
          format: date-time
          readOnly: true
    ImageResource:
      type: object
      description: Represents an image associated with a product.
      required:
        - imageUrl
        - isPrimary
      properties:
        id:
          type: integer
          readOnly: true
          example: 10
        imageUrl:
          type: string
          format: uri
          example: 'https://cdn.flexishop.com/images/product1.jpg'
        altText:
          type: string
          # Removed nullable: true
          example: 'Side view of the wireless mouse'
        isPrimary:
          type: boolean
          default: false
          example: true
        createdAt:
          type: string
          format: date-time
          readOnly: true
        updatedAt:
          type: string
          format: date-time
          readOnly: true
    SpecificationResource:
      type: object
      description: Represents a specific technical detail or feature of a product.
      required:
        - name
        - value
      properties:
        id:
          type: integer
          readOnly: true
          example: 5
        name:
          type: string
          maxLength: 255
          example: 'Battery Life'
        value:
          type: string
          maxLength: 255
          example: 'Up to 12 months'
        createdAt:
          type: string
          format: date-time
          readOnly: true
        updatedAt:
          type: string
          format: date-time
          readOnly: true

    # --- Request Bodies ---
    # Define separate request schemas for POST/PUT/PATCH
    ProductCreateRequest:
      type: object
      description: Data required to create a new product.
      required:
        - name
        - description
        - price
      properties:
        name:
          type: string
          maxLength: 255
          example: 'Ergonomic Keyboard'
        description:
          type: string
          example: 'A comfortable keyboard designed for long typing sessions.'
        price:
          type: number
          format: float
          minimum: 0.0
          example: 79.95
        status:
          type: string
          enum: [ active, inactive, discontinued ]
          default: active
          example: 'active'
        inventoryQuantity:
          type: integer
          minimum: 0
          default: 0
          example: 50
        # Use IDs for association on creation
        brandId:
          type: integer
          example: 2
        categoryIds:
          type: array
          items:
            type: integer
          example: [ 1, 2 ]
        tagIds:
          type: array
          items:
            type: integer
          example: [ 3, 4, 5 ]
    ProductUpdateRequest: # For PATCH
      type: object
      description: Fields to partially update for an existing product. All fields are optional.
      properties:
        name:
          type: string
          maxLength: 255
          example: 'Ergonomic Keyboard Pro'
        description:
          type: string
          example: 'An updated description for the pro version.'
        price:
          type: number
          format: float
          minimum: 0.0
          example: 89.95
        status:
          type: string
          enum: [ active, inactive, discontinued ]
          example: 'active'
        inventoryQuantity:
          type: integer
          minimum: 0
          example: 75
        # Allow updating associations via IDs (optional)
        brandId:
          type: integer
          # Use type: ['integer', 'null'] if explicit null is needed to *remove* association
          # example: null
        categoryIds:
          type: array
          items:
            type: integer
          description: Replace the list of associated category IDs. Send empty array `` to remove all.
        tagIds:
          type: array
          items:
            type: integer
          description: Replace the list of associated tag IDs. Send empty array `` to remove all.
...
    PageInfo:
      type: object
      description: Metadata associated with a paginated response, mirroring Spring Data Page structure.
      properties:
        totalPages:
          type: integer
          readOnly: true
        totalElements:
          type: integer
          format: int64
          readOnly: true
        size:
          type: integer
          readOnly: true
        number:
          type: integer
          readOnly: true
        numberOfElements:
          type: integer
          readOnly: true
        first:
          type: boolean
          readOnly: true
        last:
          type: boolean
          readOnly: true
        empty:
          type: boolean
          readOnly: true
        sort: # Sort is optional (not in required list)
          type: array
          items:
            $ref: '#/components/schemas/SortInfo'
          readOnly: true
      required: # Sort is not required here
        - totalPages
        - totalElements
        - size
        - number
        - numberOfElements
        - first
        - last
        - empty
    SortInfo:
      type: object
      description: Details of a single sort criterion applied to the query, mirroring Spring Data Sort structure.
      properties:
        property:
          type: string
          readOnly: true
        direction:
          type: string
          enum:
            - asc
            - desc
          readOnly: true
        ignoreCase:
          type: boolean
          readOnly: true
        nullHandling:
          type: string
          enum:
            - native
            - nulls_first
            - nulls_last
          readOnly: true
      required:
        - property
        - direction
        - ignoreCase
    ProductPage:
      description: A paginated list of products.
      allOf:
        - $ref: '#/components/schemas/PageInfo'
        - type: object
          properties:
            content:
              type: array
              items:
                $ref: '#/components/schemas/ProductResource'
          required:
            - content
...
    ProblemDetail:
      type: object
      description: A machine-readable format for specifying errors in HTTP API responses based on RFC 7807.
      properties:
        type:
          type: string
          format: uri
          default: 'about:blank'
        title:
          type: string
        status:
          type: integer
          format: int32
          minimum: 100
          maximum: 599
        detail:
          type: string
        instance:
          type: string
          format: uri-reference
        invalidParams:
          type: array
          items:
            $ref: '#/components/schemas/InvalidParam'
          # Removed nullable: true
      required:
        - type
        - title
        - status
    InvalidParam:
      type: object
      description: Represents a single validation error detail.
      properties:
        name:
          type: string
        reason:
          type: string
      required:
        - name
        - reason
  responses:
    BadRequest:
      description: Bad Request - The request syntax is malformed, contains invalid parameters, or is otherwise invalid.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetail'
          examples: # Using examples map
            validationError:
              summary: Example Validation Error
              value:
                type: '/probs/validation-error'
                title: 'Validation Error'
                status: 400
                detail: 'Request validation failed.'
                instance: '/products'
                invalidParams:
                  - name: 'name'
                    reason: 'size must be between 1 and 255'
                  - name: 'price'
                    reason: 'must be greater than or equal to 0'
...
    NotFound:
      description: Not Found - The requested resource could not be found.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetail'
          example:
            type: '/probs/not-found'
            title: 'Resource Not Found'
            status: 404
            detail: 'The product with the specified ID does not exist.'
            instance: '/products/123e4567-e89b-12d3-a456-426699999999'
    Conflict:
      description: Conflict - The request could not be completed due to a conflict with the current state of the resource.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetail'
          example:
            type: '/probs/conflict'
            title: 'Conflict'
            status: 409
            detail: 'A tag with the name "wireless" already exists.'
            instance: '/tags'
    UnprocessableEntity:
      description: Unprocessable Entity - The request was well-formed but contained semantic errors.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetail'
          example:
            type: '/probs/unprocessable-entity'
            title: 'Unprocessable Entity'
            status: 422
            detail: 'Cannot assign a discontinued category to an active product.'
            instance: '/products'
    InternalServerError:
      description: Internal Server Error - An unexpected error occurred on the server.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetail'
          example:
            type: '/probs/internal-server-error'
            title: 'Internal Server Error'
            status: 500
            detail: 'An unexpected error occurred while processing the request. Please try again later.'
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: "Authentication using a JSON Web Token (JWT) provided in the Authorization header (e.g., `Authorization: Bearer <token>`)."
security:
  - BearerAuth: